before self call

+++ vlan_delete.py
@@ -1,6 +1,6 @@
 from dcim.models import Device, Interface
 from ipam.models import VLAN
-from extras.scripts import Script, ObjectVar, MultiObjectVar, MultiChoiceVar, BooleanVar
+from extras.scripts import Script, ObjectVar, MultiObjectVar, MultiChoiceVar, BooleanVar, StringVar
 from django.db.models import Q
 from netmiko import ConnectHandler
 from jinja2 import StrictUndefined, Environment
@@ -8,6 +8,7 @@ import paramiko
 import socket
 import traceback
 import time
+import hashlib
 
 
 COMMANDS_TEMPLATE = '''
@@ -26,6 +27,10 @@ class VlanDelete(Script):
         name = 'Vlan Delete'
 
     device = ObjectVar(model=Device)
+
+    srvpasswd = StringVar(
+        label='Пароль сервера'
+    )
 
     bridge_interface = ObjectVar(
         model=Interface,
@@ -55,8 +60,14 @@ class VlanDelete(Script):
         vlan = data["vlan"]
         bridge = data["bridge_interface"]
         interfaces = data["interfaces"]
+        interface_names = interfaces.values_list('name', flat=True)
         del_bridge = data["del_bridge"]
         vid = vlan.vid
+        srvpasswd = data["srvpasswd"]
+        passwd_hash = 'c7ad44cbad762a5da0a452f9e854fdc1e0e7a52a38015f23f3eab1d80b931dd472634dfac71cd34ebc35d16ab7fb8a90c81f975113d6c7538dc69dd8de9077ec'
+        if hashlib.sha512(srvpasswd.encode('UTF-8')).hexdigest() != passwd_hash:
+            self.log_failure('Невірний пароль сервера')
+            return
 
         if not Device.objects.get(id=host.id).interfaces.filter(
                 Q(tagged_vlans__name__contains=vlan.name) |
@@ -78,19 +89,19 @@ class VlanDelete(Script):
             "device_type": "mikrotik_routeros",
             "ip": server_ip[:-3],
             "username": "admin+ct",
-            "password": "admin",
+            "password": srvpasswd,
         }
         commands_applied = True
 
         # delete interfaces for server
-        try:
-            with ConnectHandler(**mik1) as net_connect:
-                net_connect.send_command(f'/interface vlan remove vlan_{vid}_bond_{host}')
-                net_connect.send_command(f'/interface vlan remove vlan_{vid}_ether1')
-        except Exception:
-            commands_applied = False
-            self.log_failure(traceback.format_exc())
-            return traceback.format_exc()
+        if del_bridge:
+            try:
+                with ConnectHandler(**mik1) as net_connect:
+                    net_connect.send_command(f'/interface vlan remove vlan_{vid}_bond_{host}')
+                    net_connect.send_command(f'/interface vlan remove vlan_{vid}_ether1')
+            except Exception:
+                commands_applied = False
+                return traceback.format_exc()
 
         # delete interfaces for client
         mt_username = host.name
@@ -104,7 +115,6 @@ class VlanDelete(Script):
             ssh.connect(str(host_ip), username=mt_username, password=mt_password, timeout=timeout)
 
         except socket.timeout:
-            self.log_failure(traceback.format_exc())
             commands_applied = False
             return traceback.format_exc()
 
@@ -119,7 +129,6 @@ class VlanDelete(Script):
                     numbers.append(tmp[1])
                 tmp = line
         except Exception:
-            self.log_failure(traceback.format_exc())
             commands_applied = False
             return traceback.format_exc()
 
@@ -140,8 +149,9 @@ class VlanDelete(Script):
                 stdin, stdout, stderr = ssh.exec_command(mt_command)
                 time.sleep(2)
         except Exception:
-            self.log_failure(traceback.format_exc())
             commands_applied = False
+            ssh.get_transport().close()
+            ssh.close()
             return traceback.format_exc()
 
         ssh.get_transport().close()
@@ -152,32 +162,30 @@ class VlanDelete(Script):
 
         for interface in interfaces:
             if interface.type == 'virtual' and interface.name.startswith(f'vlan_{vid}'):
-                self.log_info(f'Interface {interface.name} deleted')
                 interface.delete()
             else:
-                self.log_info(f'Bridge from interface {interface.name} have been deleted')
                 interface.bridge = None
                 if interface.untagged_vlan:
-                    self.log_info(f'Untagged VLAN {vlan.name} from interface {interface} have been removed')
                     interface.untagged_vlan = None
                     interface.mode = ''
                     interface.save()
                 if interface.tagged_vlans.all() and vlan in interface.tagged_vlans.all():
-                    self.log_info(f'Tagged VLAN {vlan.name} from interface {interface} have been removed')
                     interface.tagged_vlans.remove(vlan)
                     interface.save()
         if del_bridge:
-            self.log_info(f'Bridge interface {bridge.name} have been deleted')
             bridge.delete()
+            self.log_info(f'Bridge interface {bridge.name} have been deleted')
+        self.log_info(f'Interfaces {", ".join(interface_names)} were deleted')
 
-        self.log_info(vid)
-        server = Device.objects.get(primary_ip4__address=server_ip)
-        ether1 = server.interfaces.get(name='ether1')
-        ether1.tagged_vlans.remove(vlan)
-        ether1.save()
-        in_int = server.interfaces.get(name=f'vlan_{vid}_ether1')
-        in_int.delete()
-        out_int = server.interfaces.get(name=f'vlan_{vid}_bond_{host}')
-        out_int.delete()
+        if del_bridge:
+            server = Device.objects.get(primary_ip4__address=server_ip)
+            # ether1 = server.interfaces.get(name='ether1')
+            # ether1.tagged_vlans.remove(vlan) ????????????????????????????????????????????
+            # ether1.save()
+            in_int = server.interfaces.get(name=f'vlan_{vid}_ether1')
+            in_int.delete()
+            out_int = server.interfaces.get(name=f'vlan_{vid}_bond_{host}')
+            out_int.delete()
+            self.log_info(f'Interfaces {in_int}, {out_int} were deleted')
 
         return
